% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_index.R
\docType{data}
\name{column_index}
\alias{column_index}
\title{Index File for Column Names}
\format{a `tibble` of three variables
 * `ORIGINAL_NAME`
 * `NEW_NAME`
 * `INDEX`}
\usage{
column_index
}
\description{
Index File for Column Names
}
\keyword{datasets}
