% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/filterSurvey-method.R
\name{filterSurvey}
\alias{filterSurvey}
\alias{filterSurvey,Intake24-method}
\title{Filter Surveys}
\usage{
filterSurvey(object, energy_range = c(400, 4000), min_time = 2)

\S4method{filterSurvey}{Intake24}(object, energy_range = c(400, 4000), min_time = 2)
}
\arguments{
\item{object}{an `Intake24` object}

\item{energy_range}{a numeric vector of minimum and maximum total daily energy intake (KCals). Surveys with total daily energy intake outside of the specified range will be removed}

\item{min_time}{a numeric value of minimum time to filter out}
}
\value{
object an `Intake24` object
}
\description{
Perform basic quality check on completed surveys by filtering out surveys which are extreme outlier's in total daily energy intake and have
been completed quickly (ie < 3mins)
}
