% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Intake24-class}
\alias{Intake24-class}
\title{Intake24 Class}
\description{
A S4 Class to store raw data from Intake24 Experiments
}
\section{Slots}{

\describe{
\item{\code{meta}}{a list}

\item{\code{food}}{a `tbl_df`}

\item{\code{nutrients}}{a `tbl_df`}
}}

