% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intake24_index.R
\docType{data}
\name{intake24_index}
\alias{intake24_index}
\title{Index File for Column Names}
\format{
a `tibble` of three variables
 * `Original`
 * `New`
 * `Type`
}
\usage{
intake24_index
}
\description{
Index File for Column Names
}
\keyword{datasets}
