% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openIntake24.R
\name{openIntake24}
\alias{openIntake24}
\title{Open Intake24 Experiment}
\usage{
openIntake24(input)
}
\arguments{
\item{input}{a `tibble` of the raw Intake24 data export}
}
\value{
an `Intake24` S4 object
}
\description{
Open Intake24 Experiment
}
