% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,Intake24-method}
\alias{show,Intake24-method}
\title{show-Intake24}
\usage{
\S4method{show}{Intake24}(object)
}
\arguments{
\item{object}{an `Intake24` object}
}
\description{
show-Intake24
}
