% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals-method.R
\name{totals,Intake24-method}
\alias{totals,Intake24-method}
\title{Totals}
\usage{
\S4method{totals}{Intake24}(object)
}
\arguments{
\item{object}{an `Intake24` object}
}
\value{
an `Intake24` object
}
\description{
Retrieve nutrient totals for daily intake and individual meals
}
